Added adding new pins and fixed moving them

+++ backend/ws/DTOs/Server/ServerMovesPin.cs
@@ -4,7 +4,7 @@ namespace WebsocketApi.DTOs.Server;
 
 public class ServerMovesPin : BaseDto
 {
-    public int PinId { get; set; }
+    public long PinId { get; set; }
     public int XPosition { get; set; }
     public int YPosition { get; set; }
     public string Username { get; set; }
